Basic translation from Python complete
This will be declared version 1.0
Testing shows that this version plays about twice as fast as the Python
but only has a 24% winrate.  I do not currently know why this is so weak.
I've removed the logging messages to match the flow of the original.
I will create and maintain a 'version_1' branch in line with the
structure of the original Python program.